# openapi.rapidapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ChronoKit API",
    "version": "1.0.0",
    "description": "DST-correct recurrence (RRULE), cron, and business-day computation as stateless HTTP. No accounts, no calendar platform — send a rule, get the dates.",
    "contact": {
      "name": "ChronoKit API"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://chronokit-api.chronokit-api.workers.dev",
      "description": "Production (Cloudflare Worker)"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "validation_error"
              },
              "message": {
                "type": "string",
                "example": "Request body failed validation."
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "CronNextResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Always true on success; invalid input returns a 400."
          },
          "description": {
            "type": "string",
            "description": "Present when `includeDescription` was true.",
            "example": "At 09:00 AM, Monday through Friday"
          },
          "nextRuns": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Upcoming run times as offset-aware ISO 8601 strings in `tzid`.",
            "example": [
              "2026-03-06T09:00:00.000-05:00",
              "2026-03-09T09:00:00.000-04:00"
            ]
          },
          "count": {
            "type": "integer",
            "example": 2
          }
        },
        "required": [
          "valid",
          "nextRuns",
          "count"
        ]
      },
      "CronValidateResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "example": "Invalid cron expression."
          }
        },
        "required": [
          "valid",
          "error"
        ]
      },
      "CronNextRequest": {
        "type": "object",
        "properties": {
          "expression": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "A cron expression: standard 5-field, 6-field with seconds, or a macro (@daily, @weekly, @hourly, …).",
            "example": "0 9 * * 1-5"
          },
          "tzid": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "IANA timezone identifier. Defaults to UTC when omitted.",
            "example": "America/New_York"
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "default": 5,
            "description": "Number of upcoming run times to return. Clamped to the server's per-call result cap (`MAX_RESULTS`).",
            "example": 5
          },
          "includeDescription": {
            "type": "boolean",
            "description": "Include a plain-English description of the schedule.",
            "example": true
          }
        },
        "required": [
          "expression"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/cron/next": {
      "post": {
        "tags": [
          "Cron"
        ],
        "summary": "Validate a cron expression and list its next run times",
        "description": "POST a cron expression (5-field, 6-field with seconds, or a macro like @daily). Returns validity, an optional plain-English description, and the next N run times in the requested timezone. A syntactically invalid expression returns 200 with `{ valid: false, error }` so it doubles as a validator; structural request errors return 400.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronNextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Next runs computed, OR the expression was malformed (`valid:false`).",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CronNextResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CronValidateResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Structural request error (bad body, bad types, or unknown timezone).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}